+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 
 .. note:: This version is not yet released and is under active development.
 
+* [gem]: Ignore ``default:`` prefix on package version parsing.
+
 
 `3.0.0 (2020-03-25) <https://github.com/kdeldycke/meta-package-manager/compare/v2.9.0...v3.0.0>`_
 -------------------------------------------------------------------------------------------------